+++ react/ImageListWithoutContext.tsx
@@ -6,7 +6,7 @@ import { getImagesAsJSXList } from './modules/imageAsList'
 import type { ImageListProps } from './ImageList'
 
 const ImageListWithoutContext = (props: ImageListProps) => {
-  const { images, height = 420 } = props
+  const { images, height = '420' } = props
   const { isMobile } = useDevice()
 
   const imageListContent = getImagesAsJSXList(images, isMobile, height)