+++ react/modules/imageAsList.tsx
@@ -6,12 +6,12 @@ import type { ImagesSchema } from '../ImageTypes'
 export const getImagesAsJSXList = (
   images: ImagesSchema,
   isMobile: boolean,
-  height: string | number,
+  height: number | string,
   preload?: boolean,
   experimentalPreventLayoutShift?: boolean,
   experimentalSetExplicitDimensions?: boolean
+  // eslint-disable-next-line max-params
 ) => {
-
   return images.map(
     (
       {
@@ -21,7 +21,7 @@ export const getImagesAsJSXList = (
         experimentalPreventLayoutShift: experimentalPreventLayoutShiftChild,
         experimentalSetExplicitDimensions: experimentalSetExplicitDimensionsChild,
         width = '100%',
-        height: _height,
+        height: explicitHeight,
         ...props
       },
       idx
@@ -32,7 +32,7 @@ export const getImagesAsJSXList = (
         alt={description}
         maxHeight={height}
         width={width}
-        height={_height ?? height}
+        explicitHeight={explicitHeight}
         experimentalPreventLayoutShift={
           experimentalPreventLayoutShift ?? experimentalPreventLayoutShiftChild
         }