+++ CHANGELOG.md
@@ -7,9 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-### Fixed
+### Added
 
-- fix Height on images
+- Possibility to explicit set height for images
 
 ## [0.22.3] - 2024-09-04